% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df.DISTRICTS_LARGE}
\alias{df.DISTRICTS_LARGE}
\title{{Leipzig's Bezirke (Large districts)}}
\format{
A data frame with 10.979 observations and 4 variables.
\itemize{
 \item{order }{Running number}
 \item{group }{Names of large administrative districts (Bezirke)}
 \item{lon }{Longitude}
 \item{lat }{Latitude}
}
}
\usage{
df.DISTRICTS_LARGE
}
\description{
{A dataset containing specifications to plot the boundaries
of Leipzig's 10 greater administrative districts (Bezirke)}
}
\keyword{datasets}
