% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df.DISTRICTS_SMALL}
\alias{df.DISTRICTS_SMALL}
\title{{Leipzig's Ortsteile (Small districts)}}
\format{
A data frame with 22.822 observations and 4 variables.
 \itemize{
 \item{order }{Running number}
 \item{group }{IDs of smaller local districts}
 \item{lon }{Longitude}
 \item{lat }{Latitude}
}
}
\usage{
df.DISTRICTS_SMALL
}
\description{
{A dataset containing specifications to plot the boundaries
of Leipzig's 63 smaller local districts (Ortsteile)}
}
\keyword{datasets}
