% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dfr.bundestag_2021}
\alias{dfr.bundestag_2021}
\title{{Bundestagswahlen 2021 in Leipzig}}
\format{
{data frame}
}
\usage{
dfr.bundestag_2021
}
\description{
{Bundestagswahlen 2021 in Leipzig}
}
\keyword{datasets}
