% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{get_lis_babynames}
\alias{get_lis_babynames}
\title{{Vornamen}}
\usage{
get_lis_babynames(year = 2014)
}
\arguments{
\item{year}{Number (2014 to 2019)}
}
\description{
{Leipzig's babynames between 2014 and 2015}
}
\examples{
\dontrun{
get_lis_babynames(year = 2014)
}
}
\references{
https://statistik.leipzig.de/statserv/servod.aspx
}
