% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polls.R
\name{get_lis_bundestag}
\alias{get_lis_bundestag}
\title{{Bundestagswahlen 2021 (Polls)}}
\usage{
get_lis_bundestag(...)
}
\description{
{Wahlergebnisse der Bundestagswahlen 2021 auf 3 Ebenen:}
\itemize{
\item {Stadtergebnis}
\item {Ortsteilergebnisse}
\item {Wahlkreisergebnisse}
}
}
\examples{
\dontrun{
get_lis_bundestag()
}
}
\references{
https://statistik.leipzig.de/statserv/servod.aspx
}
