% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{get_lis_edu}
\alias{get_lis_edu}
\title{{Bildung (Education)}}
\usage{
get_lis_edu(rubrik_nr = 1)
}
\arguments{
\item{rubrik_nr}{Number (1 to 5)
\itemize{
\item {1 = Allgemeinbildende Schulen}
\item {2 = Einschulungen und Nichteinschulungen}
\item {3 = Berufsbildende Schulen}
\item {4 = Hochschulen}
\item {5 = Volkshochschulen}
}}
}
\description{
{Get data from LIS about education}
}
\note{
{Only yearly data are available}
}
\examples{
\dontrun{
get_lis_edu(rubrik_nr = 1)
}
}
\references{
https://statistik.leipzig.de/statserv/servod.aspx
}
