% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{get_lis_geo}
\alias{get_lis_geo}
\title{{Stadtgebiet und Witterung (Administrative area and weather)}}
\usage{
get_lis_geo(rubrik_nr = 1)
}
\arguments{
\item{rubrik_nr}{Number (1 to 3)
\itemize{
\item {1 = Geographische Lage}
\item {2 = Flaeche und Nutzungsarten}
\item {3 = Wetterdaten}
}}
}
\description{
{Get data from LIS about geography and weather}
}
\note{
{Only yearly data are available}
}
\examples{
\dontrun{
get_lis_geo(rubrik_nr = 1)
}
}
\references{
https://statistik.leipzig.de/statserv/servod.aspx
}
