% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{get_lis_health}
\alias{get_lis_health}
\title{{Gesundheit und Soziales (Public health and social affairs)}}
\usage{
get_lis_health(rubrik_nr = 1, periode = "y")
}
\arguments{
\item{rubrik_nr}{Number (1, 2, 4)
\itemize{
\item {1 = Kindertageseinrichtungen (yearly data only)}
\item {2 = Tagespflege (yearly data only)}
\item {4 = Grundsicherung für Arbeitssuchende}
}}

\item{periode}{Yearly ('y') or quarterly ('q'). Quarterly data are nor always available.}
}
\description{
{Get data from LIS about public health and social affairs}
}
\examples{
\dontrun{
get_lis_health(rubrik_nr = 1)
}
}
\references{
https://statistik.leipzig.de/statserv/servod.aspx
}
