% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housing.R
\name{get_lis_housing}
\alias{get_lis_housing}
\title{{Bautaetigkeit und Wohnen (Building activity and housing)}}
\usage{
get_lis_housing(rubrik_nr = 1, periode = "y")
}
\arguments{
\item{rubrik_nr}{Number (1 to 5)
\itemize{
\item {1 = Wohnungsbestand (yearly data only)}
\item {2 = Wohnsituation (yearly data only)}
\item {3 = Wohnungsmieten (yearly data only)}
\item {4 = Baugenehmigungen}
\item {5 = Baufertigstellungen (yearly data only)}
}}

\item{periode}{Yearly ('y') or quarterly ('q'). Quarterly data are nor always available.}
}
\description{
{Get data from LIS about building activity and housing}
}
\examples{
\dontrun{
get_lis_housing(rubrik_nr = 1)
}
}
\references{
https://statistik.leipzig.de/statserv/servod.aspx
}
