% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kleinraeumig.R
\name{get_lis_kd}
\alias{get_lis_kd}
\title{{Kleinraeumige Daten (Leipzig's districts)}}
\usage{
get_lis_kd(kategorie_nr = 1)
}
\arguments{
\item{kategorie_nr}{Number (1 to 10 and 12)
\itemize{
\item {1 = Lage und Territorium (Geography)}
\item {2 = Bevoelkerungsbestand (Population)}
\item {3 = Bevoelkerungsbewegung (Population movement)}
\item {4 = Gesundheit und Soziales (Public health and social affairs)}
\item {5 = Bildung (Education)}
\item {6 = Bautaetigkeit und Wohnen (Building activity and housing)}
\item {7 = Erwerbstaetigkeit und Arbeitsmarkt (Occupation and labour market)}
\item {8 = Wirtschaft (Economy)}
\item {9 = Einkommen und Preise (Income and prices)}
\item {10 = Verkehr (Traffic)}
\item {12 = Oeffentliche Ordnung und Infrastruktur (Crime)}
}}
}
\description{
{Get data from LIS about Leipzig's districts}
}
\examples{
\dontrun{
get_lis_kd(kategorie_nr = 1)
}
}
\references{
https://statistik.leipzig.de/statserv/servod.aspx
}
