% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{get_lis_money}
\alias{get_lis_money}
\title{{Einkommen und Preise (Income and prices)}}
\usage{
get_lis_money(rubrik_nr = 1)
}
\arguments{
\item{rubrik_nr}{Number (1 to 2)
\itemize{
\item {1 = Lebensunterhalt}
\item {2 = Nettoeinkommen}
}}
}
\description{
{Get data from LIS about income and prices}
}
\note{
{Only yearly data are available}
}
\examples{
\dontrun{
get_lis_money(rubrik_nr = 1)
}
}
\references{
https://statistik.leipzig.de/statserv/servod.aspx
}
