% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{get_lis_pop}
\alias{get_lis_pop}
\title{{Bevoelkerungsbestand}}
\usage{
get_lis_pop(rubrik_nr = 1, periode = "y")
}
\arguments{
\item{rubrik_nr}{Number (1, 2, 5)
\itemize{
\item {1 = Einwohner}
\item {2 = Einwohner nach Alter}
\item {5 = Wohnberechtigte Einwohner nach Alter}
}}

\item{periode}{Yearly ('y') or quarterly ('q'). Quarterly data are nor always available.}
}
\description{
{Get data from LIS about the population of Leipzig}
}
\examples{
\dontrun{
get_lis_pop(rubrik_nr = 1)
}
}
\references{
https://statistik.leipzig.de/statserv/servod.aspx
}
