% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{get_lis_shapefile}
\alias{get_lis_shapefile}
\title{{Shapefile mir Grenzen der Bezirke / Ortsteile (Shapefile)}}
\usage{
get_lis_shape(shape = "Ortsteile")
}
\arguments{
\item{shape}{"Ortsteile" (default) or "Bezirke"}
}
\value{
Spatial data frame with 5 columns
\itemize{
\item {OT} {(ID of smaller local districts)}
\item {SBZ} {(ID of larger administrative districts)}
\item {Name} {(Name of smaller local /larger administrative districts)}
\item {geometry} {(Longitude, latitude etc.)}
}
}
\description{
{Shapefile with Leipzig's borders}
}
\references{
{
hhttps://opendata.leipzig.de/dataset/geo-datensatze-zu-ortsteilen-und-stadtbezirken
}
}
