% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{get_lis_streets}
\alias{get_lis_streets}
\title{{Strassennamenverzeichnis (Streets register)}}
\usage{
get_lis_streets(...)
}
\value{
Data frame with 5 columns
\itemize{
\item {STADTBEZIRKSNUMMER} {(ID of larger administrative districts)}
\item {STADTBEZIRKSNAME} {(Name of larger administrative districts)}
\item {ORTSTEIL} {(ID of smaller local district)}
\item {ORTSTEILNAME} {(Name of smaller local districts)}
\item {STRASSENNAME} {(Street name)}
}
}
\description{
{Register of Leipzig's streets}
}
\examples{
\dontrun{
get_lis_trees()
}
}
\references{
{
https://www.leipzig.de/buergerservice-und-verwaltung/unsere-stadt/gebietsgliederung-und-strassennamen/strassennamen/
}
}
