% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{get_lis_trees}
\alias{get_lis_trees}
\title{{Strassenbaumkataster (Trees, planted)}}
\usage{
get_lis_trees(...)
}
\value{
Data frame with 6 columns
\itemize{
\item {OBJECTID} {(Running number)}
\item {STRASSE_NAME} {(Name of street)}
\item {ORTSTEIL} {(Name of local district)}
\item {BAUMART_WISSENSCHAFTLICH} {(Name of tree, Latin)}
\item {BAUMART_DEUTSCH} {(Name of tree, German)}
\item {PFLANZJAHR} {(Year the tree was planted)}
}
}
\description{
{Trees planted in Leipzig}
}
\examples{
\dontrun{
get_lis_trees()
}
}
\references{
{
https://opendata.leipzig.de/dataset/strassenbaumkataster/resource/476c4909-89f8-439a-b353-4ee8d28d74d7
}
}
