% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lisr-package.r
\docType{package}
\name{lisr}
\alias{lisr-package}
\alias{lisr}
\title{lisr.}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Functions to import LIS data into R. All data will be imported as data frames.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nrkoehler/lisr}
  \item Report bugs at \url{https://github.com/nrkoehler/lisr/issues}
}

}
\author{
\strong{Maintainer}: Norbert Koehler \email{norbert.r.koehler@gmail.com} (\href{http://orcid.org/0000-0003-0724-3825}{ORCID})

}
