% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{get_lis_edu}
\alias{get_lis_edu}
\title{{Bildung}}
\usage{
get_lis_edu(rubrik_nr = 1, periode = "y")
}
\arguments{
\item{rubrik_nr}{Number (1 to 5)
\itemize{
\item {1 = Allgemeinbildende Schulen}
\item {2 = Einschulungen und Nichteinschulungen}
\item {3 = Berufsbildende Schulen}
\item {4 = Hochschulen}
\item {5 = Volkshochschulen}
}}

\item{periode}{Yearly ('y') or quarterly ('q'). Quarterly data are nor always available.}
}
\description{
{Get data from LIS about education}
}
\examples{
\dontrun{
get_lis_edu(rubrik_nr = 1)
}
}
\references{
https://statistik.leipzig.de/statserv/servod.aspx
}
