% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{get_lis_environ}
\alias{get_lis_environ}
\title{{Energie und Umwelt (Energy and environment)}}
\usage{
get_lis_environ(rubrik_nr = 1)
}
\arguments{
\item{rubrik_nr}{Number (1, 3, 4)
\itemize{
\item {1 = Immissionen}
\item {3 = Wasserhygiene}
\item {4 = Energieverbrauch}
}}
}
\description{
{Get data from LIS about energy and environment}
}
\note{
{Only yearly data are available}
}
\examples{
\dontrun{
get_lis_environ(rubrik_nr = 1)
}
}
\references{
https://statistik.leipzig.de/statserv/servod.aspx
}
