% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finances.R
\name{get_lis_finances}
\alias{get_lis_finances}
\title{{Stadtverwaltung, Kommunalpolitik, Kommunalfinanzen (Local administration, politics and finances)}}
\usage{
get_lis_finances(rubrik_nr = 1, periode = "y")
}
\arguments{
\item{rubrik_nr}{Number (1 to 10)
\itemize{
\item {1 = Personal nach Aufgabenbereichen (yearly data only)}
\item {2 = Personal nach Produktbereichen (yearly data only)}
\item {3 = Personal nach Dienstverhaeltnissen (yearly data only)}
\item {4 = Einnahmen (yearly data only)}
\item {5 = Ausgaben (yearly data only)}
\item {6 = Steuereinnahmen}
\item {7 = Realsteuervergleich (yearly data only)}
\item {8 = Zufriedenheit mit den Lebensbedingungen (yearly data only)}
\item {9 = Einsatz von Haushaltsmitteln (yearly data only)}
\item {10 = Probleme aus Buergersicht (yearly data only)}
}}

\item{periode}{Yearly ('y') or quarterly ('q'). Quarterly data are nor always available.}
}
\description{
{Get data from LIS about local administration, politics and finances}
}
\examples{
\dontrun{
get_lis_finances(rubrik_nr = 1)
}
}
\references{
https://statistik.leipzig.de/statserv/servod.aspx
}
