% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{get_lis_geo}
\alias{get_lis_geo}
\title{{Stadtgebiet und Witterung}}
\usage{
get_lis_geo(rubrik_nr = 1, periode = "y")
}
\arguments{
\item{rubrik_nr}{Number (1 to 3)
\itemize{
\item {1 = Geographische Lage}
\item {2 = Flaeche und Nutzungsarten}
\item {3 = Wetterdaten}
}}

\item{periode}{Yearly ('y') or quarterly ('q'). Quarterly data are nor always available.}
}
\description{
{Get data from LIS about geography and weather}
}
\examples{
\dontrun{
get_lis_geo(rubrik_nr = 1)
}
}
\references{
https://statistik.leipzig.de/statserv/servod.aspx
}
