% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{get_lis_move}
\alias{get_lis_move}
\title{{Bevoelkerungsbewegung (Population movement)}}
\usage{
get_lis_move(rubrik_nr = 1, periode = "y")
}
\arguments{
\item{rubrik_nr}{Number (1 to 5)
\itemize{
\item {1 = Geborene und Gestorbene (Amtliche Daten)}
\item {2 = Geborene und Gestorbene (Melderegisterdaten)}
\item {3 = Eheschliessungen und Ehescheidungen (yearly data only)}
\item {4 = Wanderungen (Amtliche Daten)}
\item {5 = Wanderungen (Melderegisterdaten)}
}}

\item{periode}{Yearly ('y') or quarterly ('q'). Quarterly data are nor always available.}
}
\description{
{Get data from LIS about population movement}
}
\examples{
\dontrun{
get_lis_move(rubrik_nr = 1)
}
}
\references{
https://statistik.leipzig.de/statserv/servod.aspx
}
