% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{get_lis_polls}
\alias{get_lis_polls}
\title{{Wahlen}}
\usage{
get_lis_polls(rubrik_nr = 1, periode = "y")
}
\arguments{
\item{rubrik_nr}{Number (1 to 4)
\itemize{
\item {1 = Europawahlen}
\item {2 = Bundestagswahlen}
\item {3 = Landtagswahlen}
\item {4 = Stadtratswahlen}
}}

\item{periode}{Yearly ('y') or quarterly ('q'). Quarterly data are nor always available.}
}
\description{
{Get data from LIS about polls}
}
\examples{
\dontrun{
get_lis_polls(rubrik_nr = 1)
}
}
\references{
https://statistik.leipzig.de/statserv/servod.aspx
}
