% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{get_lis_traffic}
\alias{get_lis_traffic}
\title{{Verkehr (Traffic and public transport)}}
\usage{
get_lis_traffic(rubrik_nr = 1, periode = "y")
}
\arguments{
\item{rubrik_nr}{Number (1 to 5)
\itemize{
\item {1 = Strassennetz}
\item {2 = Kraftfahrzeugbestand}
\item {3 = Verkehrsunfaelle}
\item {4 = Personennahverkehr}
\item {5 = Luftverkehr}
}}

\item{periode}{Yearly ('y') or quarterly ('q'). Quarterly data are nor always available.}
}
\description{
{Get data from LIS about traffic and public transport}
}
\examples{
\dontrun{
get_lis_traffic(rubrik_nr = 1)
}
}
\references{
https://statistik.leipzig.de/statserv/servod.aspx
}
